The agent api have been updated and been added to.

+++ apis/agentApi.py
@@ -1,14 +1,29 @@
 # Import modules
+import datetime
+import os, json
 from flask_restplus import Resource, Namespace, fields
-from flask import request, Flask
+from flask import request, Flask, send_from_directory, make_response
+from werkzeug.utils import secure_filename
 
 
 from Model.models import *
 from marsh import *
-
+from .authApi import user_schema, users_schema
 
 # Instance app
 app = Flask(__name__)
+
+UPLOAD_FOLDER = "./static/tours"
+app.config["UPLOAD_FOLDER"] = UPLOAD_FOLDER
+app.config["MAX_CONTENT_LENGTH"] = 16 * 1024 * 1024
+
+
+ALLOWED_EXTENSIONS = set(["jpg", "jpeg", "gif", "png"])
+
+# Check if allowed
+def allowed_file(filename):
+    return "." in filename and filename.rsplit(".", 1)[1].lower() in ALLOWED_EXTENSIONS
+
 
 # Users Namespace
 namespace2 = Namespace("agents", description="Agents related operations")
@@ -56,8 +71,7 @@ tour = namespace2.model(
         "Itinerary": fields.String,
         "Duration": fields.String,
         "StartingDate": fields.String,
-        "Special": fields.String,
-        "Price": fields.String,
+        "Price": fields.Integer,
         "Updated": fields.String,
     },
 )
@@ -93,17 +107,20 @@ class AgentInfoResource(Resource):
         return tour_schema.dump(new_agentInfo), 200
 
 
-# @namespace2.route("")
-# class AgentsResource(Resource):
-#     def get(self):
-#         """
-#         Get all Agent
-#         """
-#         agent = Users.query.filter_by(Role="Agent").all()
-#         return users_schema.dump(agent), 200
+@namespace2.route("")
+class AgentsResource(Resource):
+    def get(self):
+        """
+        Get all Agent
+        """
+        agent = Users.query.filter_by(Role=1).all()
+        if not agent:
+            return {"message": "There are no Agents!!!"}, 404
+
+        return users_schema.dump(agent), 200
 
 
-@namespace2.route("/agentId:int/tour/tourId:int/")
+@namespace2.route("/<int:agentId>/tours/<int:tourId>")
 class AgentTourResource(Resource):
     def get(self, agentId, tourId):
         """
@@ -111,7 +128,22 @@ class AgentTourResource(Resource):
         """
         # Get tour data from database
         tour = Tours.query.filter_by(UserId=agentId, TourId=tourId).first()
+        if not tour:
+            return {"message": "There are no Tours"}, 404
         # Return tour data
+        # filename = tour.TourImage
+        # try:
+        # with open(
+        #     os.path.join(app.config["UPLOAD_FOLDER"], "response.jpeg"), "rb"
+        # ) as fb:
+        #     file = fb.read()
+        #     response = make_response(file)
+        #     response.headers.set("Content-Type", "multipart/form-data")
+        #     return response
+        #     return send_from_directory("../static/tours", filename, as_attachment=False)
+
+        # except FileNotFoundError:
+        #     return {"message": "File not found"}, 404
         return tour_schema.dump(tour)
 
     @namespace2.expect(tour)
@@ -150,51 +182,78 @@ class AgentTourResource(Resource):
         """
         # Get tour data from database
         tour = Tours.query.filter_by(UserId=agentId, TourId=tourId).first()
+        if not tour:
+            return {"Message": "Tour dose not exist"}, 404
         # Delete from database
         db.session.delete(tour)
         db.session.commit()
-        return tour_schema.dump(tour), 200
+        return {"message": "Deleted"}, 200
 
 
 # SignUP a User
-@namespace2.route("/agentId:int/tours")
+@namespace2.route("/<int:agentId>/tours")
 class AgentToursResource(Resource):
     @namespace2.expect(tour)
     def post(self, agentId):
         """
         Add  Tour
         """
+        data = json.loads(request.values["data"])
+        # print(data["TourName"])
+
+        agent = Users.query.filter_by(UserId=agentId).first()
+        if agent.Role != 1:
+            return {"message": "You are not an Agent!!!"}, 400
         # Get data from api and create new instance of tour
-        new_tour = Tours(
-            TourName=request.json["TourName"],
-            TourImage=request.json["TourImage"],
-            Country=request.json["Country"],
-            Region=request.json["Region"],
-            City=request.json["City"],
-            WhatIsIncluded=request.json["WhatIsIncluded"],
-            WhatIsExcluded=request.json["WhatIsExcluded"],
-            TourDescription=request.json["TourDescription"],
-            WhatToBring=request.json["WhatToBring"],
-            Itinerary=request.json["Itinerary"],
-            Duration=request.json["Duration"],
-            StartingDate=request.json["StartingDate"],
-            Special=request.json["Special"],
-            Price=request.json["Price"],
-            Updated=request.json["Updated"],
-            UserId=agentId,
-        )
-        # Add to database
-        db.session.add(new_tour)
-        db.session.commit()
+
+        if "pic" not in request.files:
+            return {"message": "No Image File was Sent"}, 400
+
+        image = request.files["pic"]
+        if not image or not allowed_file(image.filename):
+            return {"message": "File type not allowed"}, 400
+
+        else:
+            filename = secure_filename(image.filename)
+            path = os.path.join(
+                app.config["UPLOAD_FOLDER"] + "/" + agent.UserName, filename
+            )
+            image.save(path)
+            print(filename)
+
+            new_tour = Tours(
+                TourName=data["TourName"],
+                TourImage=filename,
+                Country=data["Country"],
+                Region=data["Region"],
+                City=data["City"],
+                WhatIsIncluded=data["WhatIsIncluded"],
+                WhatIsExcluded=data["WhatIsExcluded"],
+                TourDescription=data["TourDescription"],
+                WhatToBring=data["WhatToBring"],
+                Itinerary=data["Itinerary"],
+                Duration=data["Duration"],
+                StartingDate=data["StartingDate"],
+                Price=data["Price"],
+                Updated=data["Updated"],
+                UserId=agentId,
+            )
+            # Add to database
+            db.session.add(new_tour)
+            db.session.commit()
         # return tour data
         return tour_schema.dump(new_tour), 200
+        # return "Success", 200
 
     def get(self, agentId):
-        """
+        """s
         Get All Tours
         """
         # Get every tours data from database
         tours = Tours.query.filter_by(UserId=agentId).all()
+        if not tours:
+            return {"message": "You Have no tours"}, 404
+
         # Return all tour data
         return tours_schema.dump(tours)
 
@@ -212,17 +271,21 @@ class AgentHistoryResource(Resource):
         """
 
 
-@namespace2.route("/agentId:int/history")
+@namespace2.route("/images/<imageName>")
 class AgentsHistoryResource(Resource):
-    def post(self, agentId):
+    def post(self):
         """
         Add  History
         """
 
-    def get(self, agentId):
+    def get(self, imageName):
         """
         Get Tour History
         """
+        print(imageName)
+        return send_from_directory(
+            "../static/tours/Exodes", imageName, as_attachment=False
+        )
 
 
 @namespace2.route("/agentId:int/reviews")
